pctiler/requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --extra=dev --output-file=pctiler/requirements-dev.txt ./pctiler/setup.py
#
affine==2.4.0
    # via rasterio
annotated-types==0.7.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   starlette
attrs==23.2.0
    # via
    #   cogeo-mosaic
    #   morecantile
    #   rasterio
    #   rio-tiler
boto3==1.34.123
    # via pctiler (pctiler/setup.py)
botocore==1.34.123
    # via
    #   boto3
    #   pctiler (pctiler/setup.py)
    #   s3transfer
cachetools==5.3.3
    # via
    #   cogeo-mosaic
    #   rio-tiler
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   pyproj
    #   rasterio
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   click-plugins
    #   cligj
    #   planetary-computer
    #   rasterio
click-plugins==1.1.1
    # via rasterio
cligj==0.7.2
    # via rasterio
cogeo-mosaic==7.1.0
    # via titiler-mosaic
color-operations==0.1.3
    # via rio-tiler
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
exceptiongroup==1.2.0
    # via anyio
fastapi-slim==0.111.0
    # via
    #   pctiler (pctiler/setup.py)
    #   titiler-core
fonttools==4.53.0
    # via matplotlib
geojson-pydantic==1.1.0
    # via
    #   pctiler (pctiler/setup.py)
    #   titiler-core
    #   titiler-pgstac
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via
    #   cogeo-mosaic
    #   rio-tiler
idna==3.7
    # via
    #   anyio
    #   httpx
    #   pctiler (pctiler/setup.py)
    #   requests
importlib-metadata==7.1.0
    # via rasterio
importlib-resources==6.4.0
    # via matplotlib
jinja2==3.1.4
    # via
    #   pctiler (pctiler/setup.py)
    #   titiler-core
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kiwisolver==1.4.5
    # via matplotlib
markupsafe==2.1.5
    # via jinja2
matplotlib==3.9.0
    # via pctiler (pctiler/setup.py)
morecantile==5.3.0
    # via
    #   cogeo-mosaic
    #   rio-tiler
    #   supermorecado
    #   titiler-core
numexpr==2.9.0
    # via rio-tiler
numpy==1.26.4
    # via
    #   color-operations
    #   contourpy
    #   matplotlib
    #   numexpr
    #   rasterio
    #   rio-tiler
    #   shapely
    #   snuggs
    #   titiler-core
orjson==3.10.4
    # via pctiler (pctiler/setup.py)
packaging==24.1
    # via
    #   matplotlib
    #   planetary-computer
pillow==10.3.0
    # via
    #   matplotlib
    #   pctiler (pctiler/setup.py)
planetary-computer==1.0.0
    # via pctiler (pctiler/setup.py)
psycopg[binary,pool]==3.1.18
    # via pctiler (pctiler/setup.py)
psycopg-binary==3.1.18
    # via psycopg
psycopg-pool==3.2.1
    # via psycopg
pydantic==2.7.4
    # via
    #   cogeo-mosaic
    #   fastapi-slim
    #   geojson-pydantic
    #   morecantile
    #   pctiler (pctiler/setup.py)
    #   planetary-computer
    #   pydantic-settings
    #   rio-tiler
    #   titiler-core
    #   titiler-pgstac
pydantic-core==2.18.4
    # via pydantic
pydantic-settings==2.3.3
    # via
    #   cogeo-mosaic
    #   titiler-pgstac
pyparsing==3.1.2
    # via
    #   matplotlib
    #   snuggs
pyproj==3.6.1
    # via morecantile
pystac==1.10.1
    # via
    #   pctiler (pctiler/setup.py)
    #   planetary-computer
    #   pystac-client
    #   rio-tiler
pystac-client==0.6.1
    # via planetary-computer
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   matplotlib
    #   pystac
    #   pystac-client
python-dotenv==1.0.1
    # via
    #   planetary-computer
    #   pydantic-settings
pytz==2024.1
    # via planetary-computer
rasterio==1.3.10
    # via
    #   cogeo-mosaic
    #   pctiler (pctiler/setup.py)
    #   rio-tiler
    #   supermorecado
    #   titiler-core
requests==2.32.3
    # via
    #   pctiler (pctiler/setup.py)
    #   planetary-computer
    #   pystac-client
rio-tiler==6.6.1
    # via
    #   cogeo-mosaic
    #   titiler-core
s3transfer==0.10.1
    # via boto3
shapely==2.0.3
    # via cogeo-mosaic
simplejson==3.19.2
    # via titiler-core
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
snuggs==1.4.7
    # via rasterio
starlette==0.37.2
    # via fastapi-slim
supermorecado==0.1.2
    # via cogeo-mosaic
titiler-core==0.18.3
    # via
    #   pctiler (pctiler/setup.py)
    #   titiler-mosaic
    #   titiler-pgstac
titiler-mosaic==0.18.3
    # via
    #   pctiler (pctiler/setup.py)
    #   titiler-pgstac
titiler-pgstac==1.3.0
    # via pctiler (pctiler/setup.py)
types-requests==2.31.0.6
    # via pctiler (pctiler/setup.py)
types-urllib3==1.26.25.14
    # via types-requests
typing-extensions==4.10.0
    # via
    #   anyio
    #   fastapi-slim
    #   psycopg
    #   psycopg-pool
    #   pydantic
    #   pydantic-core
    #   starlette
    #   titiler-core
urllib3==1.26.18
    # via
    #   botocore
    #   requests
zipp==3.19.2
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
